+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 ## CHANGELOG
 
-### UNRELEASED CHANGES
+### 4.0.0
+- BREAKING CHANGE: Start app by fetching auth document first instead of catalog root. This means that the `libraries` key in the config file needs to be set with auth document urls instead of catalog root urls.
 - Feature: Support auth-protected collections.
-- BREAKING CHANGE: Start app by fetching auth document first instead of catalog root. This means that config file needs to be set with auth document urls now.
 
 ### 3.1.3
 - Fix: Update the production docker publish Github Action to v2, allowing for use of private Github package registry for AxisNow Decryptor.